refactor catalog-item.tsx and extract catalog tile, refactor cell.tsx to use padding instead of subtracting the height width to create spacing
fix unit tests
