+++ frontend/packages/console-shared/src/components/virtualized-grid/Cell.tsx
@@ -1,18 +1,25 @@
 import * as React from 'react';
 import { CellMeasurer } from 'react-virtualized';
 import { RenderCell, RenderHeader, GridChildrenProps, Item, CellItem } from './types';
-import { getHeightAndWidthOfCell, CellMeasurementContext } from './utils';
+import { CellMeasurementContext } from './utils';
 
 type CellProps = {
   renderCell: RenderCell;
   renderHeader?: RenderHeader;
 } & GridChildrenProps;
 
-const Cell: React.FC<CellProps> = ({ data, columnCount, items, renderCell, renderHeader }) => {
+const Cell: React.FC<CellProps> = ({
+  data,
+  columnCount,
+  items,
+  rowCount,
+  renderCell,
+  renderHeader,
+}) => {
   const { cache, cellMargin } = React.useContext(CellMeasurementContext);
   const {
     key,
-    style: { height: cellHeight, width: cellWidth, left, top, ...style },
+    style: { width, ...style },
     columnIndex,
     rowIndex,
     parent,
@@ -21,14 +28,14 @@ const Cell: React.FC<CellProps> = ({ data, columnCount, items, renderCell, rende
   const item: CellItem = items[index];
   const isItemString = typeof item === 'string';
 
-  const { width, height } = getHeightAndWidthOfCell(cellHeight, cellWidth, item);
-  const wrapperStyles = {
+  const cellStyle = {
     ...style,
-    height,
-    width,
-    left: Number(left) + cellMargin,
-    top: Number(top) + cellMargin,
+    width: isItemString ? '100%' : width,
+    padding: `${cellMargin}px ${cellMargin}px ${
+      rowIndex === rowCount - 1 ? `${cellMargin}px` : 0
+    } ${cellMargin}px`,
   };
+
   return item ? (
     <CellMeasurer
       cache={cache}
@@ -37,7 +44,7 @@ const Cell: React.FC<CellProps> = ({ data, columnCount, items, renderCell, rende
       parent={parent}
       rowIndex={rowIndex}
     >
-      <div style={wrapperStyles}>
+      <div style={cellStyle}>
         {isItemString ? renderHeader(item as string) : renderCell(item as Item)}
       </div>
     </CellMeasurer>
