+++ frontend/packages/console-shared/src/components/virtualized-grid/Grid.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import { Grid as GridComponent, GridCellProps } from 'react-virtualized';
-import { CELL_PADDING } from './const';
 import { Item, GridChildrenProps } from './types';
 import { CellMeasurementContext } from './utils';
 import './Grid.scss';
@@ -14,7 +13,7 @@ type GridProps = {
 };
 
 const Grid: React.FC<GridProps> = ({ height, width, scrollTop, items, children }) => {
-  const { cache, cellWidth, cellMargin, overscanRowCount } = React.useContext(
+  const { cache, cellWidth, cellMargin, overscanRowCount, estimatedCellHeight } = React.useContext(
     CellMeasurementContext,
   );
   const itemCount = items.length;
@@ -22,7 +21,7 @@ const Grid: React.FC<GridProps> = ({ height, width, scrollTop, items, children }
   const columnCountEstimate = Math.max(1, Math.floor(width / idealItemWidth));
   const rowCount = Math.ceil(itemCount / columnCountEstimate);
   const columnCount = Math.max(1, itemCount && Math.ceil(itemCount / rowCount));
-  const cellRenderer = (data: GridCellProps) => children({ data, columnCount, items });
+  const cellRenderer = (data: GridCellProps) => children({ data, columnCount, items, rowCount });
   return (
     <GridComponent
       className="ocs-grid"
@@ -30,7 +29,8 @@ const Grid: React.FC<GridProps> = ({ height, width, scrollTop, items, children }
       height={height ?? 0}
       width={width}
       scrollTop={scrollTop}
-      rowHeight={(params) => cache.rowHeight(params) + CELL_PADDING}
+      rowHeight={(params) => cache.rowHeight(params)}
+      estimatedRowSize={estimatedCellHeight}
       deferredMeasurementCache={cache}
       columnWidth={idealItemWidth}
       rowCount={rowCount}
